refactor to have the async wrapper here
