+++ pkg/awsds/utils.go
@@ -6,7 +6,9 @@ import (
 	"runtime"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/build"
+	"github.com/grafana/grafana-plugin-sdk-go/data"
 )
 
 // GetUserAgentString returns an agent that can be parsed in server logs
@@ -36,4 +38,24 @@ func GetUserAgentString(name string) string {
 		buildInfo.Version,
 		buildInfo.Hash,
 		grafanaVersion)
+}
+
+func getDatasourceUID(settings backend.DataSourceInstanceSettings) string {
+	datasourceUID := settings.UID
+	// Grafana < 8.0 won't include the UID yet
+	if datasourceUID == "" {
+		datasourceUID = fmt.Sprintf("%d", settings.ID)
+	}
+	return datasourceUID
+}
+
+// getErrorFrameFromQuery returns a error frames with empty data and meta fields
+func getErrorFrameFromQuery(query *AsyncQuery) data.Frames {
+	frames := data.Frames{}
+	frame := data.NewFrame(query.RefID)
+	frame.Meta = &data.FrameMeta{
+		ExecutedQueryString: query.RawSQL,
+	}
+	frames = append(frames, frame)
+	return frames
 }
