+++ pkg/sql/datasource/utils.go
@@ -4,10 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/grafana/grafana-aws-sdk/pkg/sql/api"
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/resource/httpadapter"
-	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/grafana/sqlds/v2"
 )
 
@@ -21,24 +18,4 @@ func GetDatasourceID(ctx context.Context) int64 {
 		return plugin.DataSourceInstanceSettings.ID
 	}
 	return 0
-}
-
-func getDatasourceUID(settings backend.DataSourceInstanceSettings) string {
-	datasourceUID := settings.UID
-	// Grafana < 8.0 won't include the UID yet
-	if datasourceUID == "" {
-		datasourceUID = fmt.Sprintf("%d", settings.ID)
-	}
-	return datasourceUID
-}
-
-// getErrorFrameFromQuery returns a error frames with empty data and meta fields
-func getErrorFrameFromQuery(query *api.AsyncQuery) data.Frames {
-	frames := data.Frames{}
-	frame := data.NewFrame(query.RefID)
-	frame.Meta = &data.FrameMeta{
-		ExecutedQueryString: query.RawSQL,
-	}
-	frames = append(frames, frame)
-	return frames
 }
