+++ pkg/sql/driver/async/driver.go
@@ -1,13 +1,14 @@
 package async
 
 import (
+	"github.com/grafana/grafana-aws-sdk/pkg/awsds"
 	"github.com/grafana/grafana-aws-sdk/pkg/sql/api"
 	sqlDriver "github.com/grafana/grafana-aws-sdk/pkg/sql/driver"
 )
 
 type Driver interface {
 	sqlDriver.Driver
-	GetAsyncDB() (api.AsyncDB, error)
+	GetAsyncDB() (awsds.AsyncDB, error)
 }
 
 type Loader func(api.AWSAPI) (Driver, error)
